Revert "avanzando con hibernate"
This reverts commit 0d089be8eee098a59518c77e68200577d025a2f6.

+++ interfaces/src/main/java/ar/edu/itba/interfaces/UserDao.java
@@ -4,15 +4,15 @@ import ar.edu.itba.model.User;
 
 public interface UserDao {
 	
-	public User findById(int id);
+	public User findById(long id);
 	
 	public User create(String username, String password, String email);
 	
-	public String getUsername(int id);
+	public String getUsername(long id);
 
 	public User findByUsername(String username);
 
-	public String getEmail(int id);
+	public String getEmail(long id);
 	
 	public String getEmail(String username);
